+++ lib/enum/enum_nav_ani.dart
@@ -9,34 +9,35 @@ enum NavAni { Left, Right, Top, Bottom, Fade, Ripple, Blink }
 
 extension NavAniExt on NavAni {
   PageRoute<T> createRoute<T>(
-      Widget screen, BuildContext context, int durationMs) {
+      Widget screen, BuildContext? context, int durationMs) {
     switch (this) {
       case NavAni.Left:
-        return SlideFromLeftRoute(widget: screen, durationMs: durationMs);
+        return SlideFromLeftRoute(screen, durationMs: durationMs);
       case NavAni.Right:
         return Nav.getPushRightRoute(screen,
-            context: context, prohibitSwipeBack: false, durationMs: durationMs);
+            context: context,
+            prohibitSwipeBack: false,
+            durationMs: durationMs) as PageRoute<T>;
       case NavAni.Top:
-        return SlideFromTopRoute(widget: screen, durationMs: durationMs);
+        return SlideFromTopRoute(screen, durationMs: durationMs);
       case NavAni.Bottom:
-        return SlideFromBottomRoute(widget: screen, durationMs: durationMs);
+        return SlideFromBottomRoute(screen, durationMs: durationMs);
       case NavAni.Fade:
-        return FadeRoute(widget: screen, durationMs: durationMs);
+        return FadeRoute(screen, durationMs: durationMs);
       case NavAni.Blink:
-        return BlinkRoute(widget: screen);
+        return BlinkRoute(screen);
       case NavAni.Ripple:
         if (Nav.height == null) {
-          Nav.initDeviceSize(context);
+          Nav.initDeviceSize(context!);
         }
-        return RoundRevealRoute(
-            widget: screen,
-            maxRadius: Nav.height + Nav.width / 2,
+        return RoundRevealRoute(screen,
+            maxRadius: Nav.height! + Nav.width / 2,
             centerAlignment: Alignment.bottomRight,
             centerOffset: Offset(10, 10),
             minRadius: 10,
             durationMs: durationMs);
       default:
-        return SlideFromRightRoute(widget: screen);
+        return SlideFromRightRoute(screen);
     }
   }
 }